pyproject.toml
[project]
name = "voxelengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "freetype-py>=2.5.1",
    "glfw>=2.10.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pyglm>=2.8.3",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
]
